gtm-keras-h5-predictor/gtm-keras-h5-predictor/requirements.txt
tensorflow==2.9.1  # 
gradio==4.29.0  # 
Pillow==9.1.0  # 
numpy==1.22.3  # 


